requirements.txt
kafka-python
flask
waitress